test_requirements.txt
django-nose==1.4.6
pint==0.12
nose==1.3.7
pinocchio==0.4.2
mock==4.0.2
tox==3.15.1
